#2465 r23269 - r23272, r23274 - r23275, r23277 - r23278 をマージ

+++ data/class/pages/admin/contents/LC_Page_Admin_Contents_FileManager.php
@@ -126,10 +126,13 @@ class LC_Page_Admin_Contents_FileManager extends LC_Page_Admin_Ex
                         $disp_error = '※ ディレクトリをダウンロードすることは出来ません。<br/>';
                         $this->setDispError('select_file', $disp_error);
                     } else {
-                        // ファイルダウンロード
-                        $objFileManager->sfDownloadFile($objFormParam->getValue('select_file'));
-                        SC_Response_Ex::actionExit();
-                    }
+                       $path_exists = SC_Utils::checkFileExistsWithInBasePath($objFormParam->getValue('select_file'),USER_REALDIR);
+                       if ($path_exists) {
+                         // ファイルダウンロード
+                         $objFileManager->sfDownloadFile($objFormParam->getValue('select_file'));
+                         SC_Response_Ex::actionExit();
+                      }
+                  }
                 }
                 break;
             // ファイル削除
@@ -138,9 +141,9 @@ class LC_Page_Admin_Contents_FileManager extends LC_Page_Admin_Ex
                 $this->lfInitParamModeView($objFormParam);
                 $objFormParam->setParam($this->createSetParam($_POST));
                 $objFormParam->convParam();
-
                 $this->arrErr = $objFormParam->checkError();
-                if (SC_Utils_Ex::isBlank($this->arrErr)) {
+                $path_exists = SC_Utils::checkFileExistsWithInBasePath($objFormParam->getValue('select_file'),USER_REALDIR);
+                if (SC_Utils_Ex::isBlank($this->arrErr) && ($path_exists)) {
                     $objFileManager->deleteFile($objFormParam->getValue('select_file'));
                 }
                 break;
